[TH-513] 피드백 상세
- 피드백 상세 작업
- 피드백 화면 매니페스트 등록

+++ core/network/src/main/java/com/threedollar/network/data/store/Feedback.kt
@@ -16,5 +16,5 @@ data class Feedback(
     val count: Int? = 0,
 
     @SerializedName("ratio")
-    val ratio: Double? = 0.0
+    val ratio: Float? = 0.0f
 )
